requirements.txt
numpy~=2.2.4
pandas~=2.2.3
matplotlib~=3.9.2
seaborn~=0.13.2
nltk~=3.9.1
scikit-learn~=1.5.2
wordcloud~=1.9.4
textblob~=0.18.0.post0

transformers[torch]
datasets
rich
evaluate
peft
torchmetrics
adapters
graphviz
torchview
bitsandbytes

black
unidecode
langdetect
transformers
accelerate
fbgemm-gpu
